+++ src/components/Card/components/CardIconsWrapper/types.ts
@@ -1,0 +1,11 @@
+import type {
+    FC,
+    HTMLAttributes,
+    ReactNode
+} from 'react'
+
+export interface CardIconsWrapperProps extends HTMLAttributes<HTMLDivElement> {
+    children?: ReactNode
+}
+
+export type CardIconsWrapperComponent = FC<CardIconsWrapperProps>