+++ src/components/Card/components/CardImage/CardImage.tsx
@@ -1,0 +1,41 @@
+import { useCard } from '@components/Card/contexts/CardContext'
+
+import Skeleton from '@components/Skeleton'
+
+import * as Styled from './style'
+import type * as Types from './types'
+import type { SkeletonVariantType } from '@components/Skeleton/types'
+
+const CardImage: Types.CardImageComponent = (props) => {
+    const {
+        style,
+        className,
+        ...DOMProps
+    } = props
+
+    const {
+        loading,
+        variant,
+        vertical,
+        compact
+    } = useCard()
+
+    return (
+        <Styled.Wrapper
+            style={style}
+            className={className}
+            $vertical={vertical}
+            $compact={compact}
+        >
+            {
+                loading ? (
+                    <Skeleton variant={variant as SkeletonVariantType}/>
+                ) : (
+                    <Styled.Image {...DOMProps}/>
+                )
+            }
+        </Styled.Wrapper>
+    )
+}
+
+export default CardImage