+++ src/components/Card/components/CardParagraph/types.ts
@@ -1,0 +1,10 @@
+import type {
+    FC,
+    HTMLAttributes
+} from 'react'
+
+export interface CardParagraphProps extends HTMLAttributes<HTMLParagraphElement> {
+    text: string
+}
+
+export type CardParagraphComponent = FC<CardParagraphProps>