+++ src/components/Card/components/CardTitle/types.ts
@@ -1,0 +1,10 @@
+import type {
+    FC,
+    HTMLAttributes
+} from 'react'
+
+export interface CardTitleProps extends HTMLAttributes<HTMLParagraphElement> {
+    text: string
+}
+
+export type CardTitleComponent = FC<CardTitleProps>