+++ src/components/Portal/Portal.test.tsx
@@ -1,0 +1,12 @@
+import Portal from './index'
+import { render } from '@testing-library/react'
+
+describe('<Portal />', () => {
+    it('should not crash', () => {
+        render(
+            <Portal>
+                {'Henllo!'}
+            </Portal>
+        )
+    })
+})