+++ src/components/Portal/types.ts
@@ -1,0 +1,12 @@
+import {
+    ReactNode,
+    FC
+} from 'react'
+
+export type PortalProps = {
+    dom?: Element | DocumentFragment,
+    portalWhenMounted?: boolean,
+    children: ReactNode
+}
+
+export type PortalComponent = FC<PortalProps>