+++ src/contexts/ThemeContext/index.ts
@@ -1,1 +1,1 @@
-export { default } from './ThemeContext'
+export * from './ThemeContext'