+++ src/hooks/useTheme/useTheme.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react"
 
-import ThemeContext from "@root/contexts/ThemeContext"
+import { ThemeContext } from "@root/contexts/ThemeContext"
 
 const useTheme = () => useContext(ThemeContext)