+++ src/index.ts
@@ -1,11 +1,56 @@
-export * from '@components/Alert'
-export { default as Alert } from '@components/Alert'
+// Utils ====>
+
+export {
+    theme,
+    getTheme,
+} from '@utils/theme'
 
 export * from '@components/Button'
 export { default as Button } from '@components/Button'
 
-export * from '@components/ButtonGroup'
-export { default as ButtonGroup } from '@components/ButtonGroup'
+// <==== Utils
+
+
+// Contexts ====>
+
+export {
+    ThemeContext,
+    ThemeProvider
+} from '@root/contexts/ThemeContext'
+
+// <==== Contexts
+
+
+// Hooks ====>
+
+export * from '@hooks/useDeepCompareCallback'
+export { default as useDeepCompareCallback } from '@hooks/useDeepCompareCallback'
+
+export * from '@hooks/useDeepCompareEffect'
+export { default as useDeepCompareEffect } from '@hooks/useDeepCompareEffect'
+
+export * from '@hooks/useDeepCompareMemoize'
+export { default as useDeepCompareMemoize } from '@hooks/useDeepCompareMemoize'
+
+export * from '@hooks/useLazyRef'
+export { default as useLazyRef } from '@hooks/useLazyRef'
+
+export * from '@hooks/useLifeCycle'
+export { default as useLifeCycle } from '@hooks/useLifeCycle'
+
+export * from '@hooks/useTheme'
+export { default as useTheme } from '@hooks/useTheme'
+
+export * from '@hooks/useScroll'
+export { default as useScroll } from '@hooks/useScroll'
+
+export * from '@hooks/useMounted'
+export { default as useMounted } from '@hooks/useMounted'
+
+// <==== Hooks
+
+
+// Components ====>
 
 export * from '@components/Checkbox'
 export { default as Checkbox } from '@components/Checkbox'
@@ -37,26 +82,16 @@ export { default as ThemeContext } from '@contexts/ThemeContext'
 export * from '@hooks/useDeepCompareCallback'
 export { default as useDeepCompareCallback } from '@hooks/useDeepCompareCallback'
 
-export * from '@hooks/useDeepCompareEffect'
-export { default as useDeepCompareEffect } from '@hooks/useDeepCompareEffect'
+export * from '@components/ButtonGroup'
+export { default as ButtonGroup } from '@components/ButtonGroup'
 
-export * from '@hooks/useDeepCompareMemoize'
-export { default as useDeepCompareMemoize } from '@hooks/useDeepCompareMemoize'
+export * from '@components/Skeleton'
+export { default as Skeleton } from '@components/Skeleton'
 
-export * from '@hooks/useLazyRef'
-export { default as useLazyRef } from '@hooks/useLazyRef'
+export * from '@components/Card'
+export { default as Card } from '@components/Card'
 
-export * from '@hooks/useLifeCycle'
-export { default as useLifeCycle } from '@hooks/useLifeCycle'
+export * from '@components/Portal'
+export { default as Portal } from '@components/Portal'
 
-export * from '@hooks/useTheme'
-export { default as useTheme } from '@hooks/useTheme'
-
-export * from '@utils/deepCompareEquals'
-export { default as deepCompareEquals } from '@utils/deepCompareEquals'
-
-export * from '@utils/environment'
-export { default as environment } from '@utils/environment'
-
-export * from '@utils/theme'
-export { default as theme } from '@utils/theme'
+// <==== Components